LayoutTests/fast/dom/DeviceOrientation/window-property-expected.txt

Tests that the window.ondeviceorientation property is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof window.ondeviceorientation == 'object' is true
PASS hasOnDeviceOrientationProperty() is true
PASS 'ondeviceorientation' in window is true
PASS window.hasOwnProperty('ondeviceorientation') is true
PASS successfullyParsed is true

TEST COMPLETE
